refactor AppDetail Usage Paging Chart.

+++ app/src/main/java/com/chs/yourapphistory/presentation/screen/common/BarArea.kt
@@ -2,7 +2,7 @@ package com.chs.yourapphistory.presentation.screen.common
 
 data class BarArea(
     val idx: Int,
-    val value: Long,
+    val value: Int,
     val xStart: Float,
     val xEnd: Float
 )
